Labels & folders migration + (almost) latest Nylas

+++ inbox/models/base.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer
+from sqlalchemy import Column, BigInteger
 from sqlalchemy.ext.declarative import as_declarative, declared_attr
 
 from inbox.models.mixins import AutoTimestampMixin
@@ -8,8 +8,9 @@ from inbox.models.mixins import AutoTimestampMixin
 class MailSyncBase(AutoTimestampMixin):
     """
     Provides automated table name, primary key column, and audit timestamps.
+
     """
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
 
     @declared_attr
     def __tablename__(cls):
